+++ pennylane_lightning/core/src/simulators/lightning_qubit/gates/tests/Test_GateImplementations_Nonparam.cpp
@@ -31,7 +31,7 @@
  * @file Test_GateImplementations_Nonparam.cpp
  *
  * This file contains tests for non-parameterized gates. List of such gates are
- * [PauliX, PauliY, PauliZ, Hadamard, S, T, CNOT, SWAP, CZ, Toffoli, CSWAP].
+ * [PauliX, PauliY, PauliZ, Hadamard, S, SX, T, CNOT, SWAP, CZ, Toffoli, CSWAP].
  */
 
 /// @cond DEV
@@ -243,6 +243,30 @@ template <typename PrecisionT, class GateImplementation> void testApplyT() {
     }
 }
 PENNYLANE_RUN_TEST(T);
+
+template <typename PrecisionT, class GateImplementation> void testApplySX() {
+    // Test using |000> state
+    using ComplexT = std::complex<PrecisionT>;
+    const size_t num_qubits = 3;
+
+    constexpr ComplexT z(0.0, 0.0);
+    constexpr ComplexT p(0.5, 0.5);
+    constexpr ComplexT m(0.5, -0.5);
+
+    const std::vector<std::vector<ComplexT>> expected_results = {
+        {p, z, z, z, m, z, z, z},
+        {p, z, m, z, z, z, z, z},
+        {p, m, z, z, z, z, z, z}};
+
+    for (size_t index = 0; index < num_qubits; index++) {
+        auto st = createZeroState<ComplexT>(num_qubits);
+
+        GateImplementation::applySX(st.data(), num_qubits, {index}, false);
+
+        CHECK(st == approx(expected_results[index]));
+    }
+}
+PENNYLANE_RUN_TEST(SX);
 /*******************************************************************************
  * Two-qubit gates
  ******************************************************************************/
@@ -816,6 +840,28 @@ TEMPLATE_TEST_CASE("StateVectorLQubitManaged::applyOperation non-param "
                                       std::vector<bool>{true},
                                       std::vector<std::size_t>{wire});
             sv1.applyOperation("T", std::vector<std::size_t>{control},
+                               std::vector<bool>{true},
+                               std::vector<std::size_t>{wire});
+            REQUIRE(sv0.getDataVector() ==
+                    approx(sv1.getDataVector()).margin(margin));
+        }
+    }
+    DYNAMIC_SECTION("N-controlled SX - "
+                    << "controls = {" << control << "} "
+                    << ", wires = {" << wire << "} - "
+                    << PrecisionToName<PrecisionT>::value) {
+        if (control != wire) {
+            auto st0 = createRandomStateVectorData<PrecisionT>(re, num_qubits);
+            sv0.updateData(st0);
+            sv1.updateData(st0);
+
+            const std::vector<std::complex<PrecisionT>> matrix =
+                getSX<std::complex, PrecisionT>();
+
+            sv0.applyControlledMatrix(matrix, std::vector<std::size_t>{control},
+                                      std::vector<bool>{true},
+                                      std::vector<std::size_t>{wire});
+            sv1.applyOperation("SX", std::vector<std::size_t>{control},
                                std::vector<bool>{true},
                                std::vector<std::size_t>{wire});
             REQUIRE(sv0.getDataVector() ==